add common game component, modify addition and subtraction components
