+++ src/games/math/_common/constants/operations.js
