+++ src/games/math/_common/digitsPanel/digitsPanel.css
