+++ src/games/math/_common/game/game.css
@@ -1,0 +1,7 @@
+.addition {
+  background: gold;
+}
+
+.subtraction {
+  background: pink;
+}
