+++ src/games/math/_common/gameDisplay/digit.js
