+++ src/games/math/_common/gameDisplay/operation.css
