+++ src/games/math/_common/gameDisplay/operation.js
