+++ src/games/math/_common/gameDisplay/total.css
