+++ src/games/math/_common/gameDisplay/total.js
