+++ src/games/math/_common/statusBar/levelDisplay.js
