+++ src/games/math/_common/statusBar/scoreDisplay.js
