+++ src/games/math/_common/statusBar/sessionTimer.css
