+++ src/games/math/_common/statusBar/solutionDisplay.css
