+++ src/games/math/_common/statusBar/statusBar.css
