+++ src/games/math/_common/statusBar/statusBar.js
