--- src/games/math/addition/addition.css
@@ -1,3 +1,0 @@
-.addition {
-  background: gold;
-}
