--- src/games/math/subtraction/subtraction.css
@@ -1,3 +1,0 @@
-.subtraction {
-  background: purple;
-}
